ability to specify custom TopologyProvider in TopologyRefresher
